LinearRegression and first model which search koeff

+++ MLs/Components/Matrix2x2.cs
@@ -6,7 +6,7 @@ namespace MLs.Components
         public double M11; public double M12;
         public double M21; public double M22;
 
-        public Matrix2x2(float m11, float m12, float m21, float m22)
+        public Matrix2x2(double m11, double m12, double m21, double m22)
         {
             M11 = m11; M12 = m12;
             M21 = m21; M22 = m22;
